Move functionality from Artifact class to ArtifactCommand class.
